ensure elapsed time is stored to testunit
When executing outside the context of Go testing tool, we were not
marking a test unit finished upon successful completion of the test
spec, resulting in elapsed times of 0 when run with the gdt CLI tool.
This fixes that by ensuring we set the elapsed time by calling Finish()
on the test unit.

Signed-off-by: Jay Pipes <jaypipes@gmail.com>